requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pylightning==0.0.7.3
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
